Fix data rank; Remove board-extra
