+++ src/utils/const.js
@@ -1,0 +1,5 @@
+export const RANK_USER = [
+  `novice`,
+  `fan`,
+  `movie buff`
+];
