--- src/view/board-extra.js
@@ -1,8 +1,0 @@
-export const createBoardExtraTemplate = () => {
-  return `
-    <section class="films-list films-list--extra">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-      <div class="films-list__container"></div>
-    </section>
-  `;
-};
